added new models

+++ src/main/java/br/terracore/server/address/Address.java
@@ -14,7 +14,7 @@ import javax.xml.bind.annotation.XmlRootElement;
 public class Address implements Serializable {
         @Id
         @GeneratedValue
-        private Integer id;
+        private Long id;
         
         private String  name;
         
@@ -41,7 +41,7 @@ public class Address implements Serializable {
         }
         
         public Address(
-                       Integer id,
+                       Long id,
                        String name,
                        String number,
                        String extra,
@@ -64,11 +64,11 @@ public class Address implements Serializable {
                 this.featureId = featureId;
         }
         
-        public Integer getId() {
+        public Long getId() {
                 return id;
         }
         
-        public void setId(Integer id) {
+        public void setId(Long id) {
                 this.id = id;
         }
